Refactor `response.py`
Move all classes from `response.py` to `base.py`.
