+++ server/app/interfaces/registry.py
@@ -15,8 +15,7 @@ from werkzeug.wrappers import Request, Response
 import server.app.server_model as server_model
 from basyx.aas import model
 from server.app.api_utils.http_api_helpers import HTTPApiDecoder, Base64URLConverter, is_stripped_request
-from server.app.api_utils.response import APIResponse
-from server.app.interfaces.base import ObjectStoreWSGIApp
+from server.app.interfaces.base import ObjectStoreWSGIApp, APIResponse
 
 
 class RegistryAPI(ObjectStoreWSGIApp):
